+++ examples/basic/main.rs
@@ -37,9 +37,9 @@ fn main() {
     context.insert("bio", &"<script>alert('pwnd');</script>");
 
     // A one off template
-    Tera::one_off("hello", &Context::new(), true).unwrap();
+    Tera::one_off("hello", Context::new(), true).unwrap();
 
-    match TEMPLATES.render("users/profile.html", &context) {
+    match TEMPLATES.render("users/profile.html", context) {
         Ok(s) => println!("{:?}", s),
         Err(e) => {
             println!("Error: {}", e);