+++ CHANGELOG.md
@@ -4,10 +4,13 @@
 
 ### Breaking changes
 
-- Now requires Rust 1.30
+- Now requires Rust 1.31
 - Removed error-chain errors and added rich Error enum instead
 - Filter, Tester and Function are now traits and now take borrowed values instead of owned
 - Updated for 2018 edition
+- Split a `render/one_off` with `render_borrowed/one_off_borrowed`. `render/one_off` now takes
+a `Context` by value to avoid cloning.
+
 
 ### Others